# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SOY MONEY OS API
servers:
  - url: /api
    description: Base API path
...
security:
  - ClerkOwner: []
paths:
...
  /discovery/research:
    post:
      operationId: researchDiscovery
      tags: [discovery]
      summary: Research a category using free public sources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoveryResearchInput"
      responses:
        "201":
          description: Research completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryResearchResponse"
        "422":
          description: No corroborated opportunity
    get:
      operationId: listDiscoveryResearch
      tags: [discovery]
      summary: List durable discovery research runs
      responses:
        "200":
          description: Research runs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DiscoveryResearchRun" }
  /discovery/research/{id}:
    get:
      operationId: getDiscoveryResearch
      tags: [discovery]
      summary: Get a research run and its findings
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Research run and findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryResearchResponse"
...
components:
  securitySchemes:
    ClerkOwner:
      type: apiKey
      in: cookie
      name: __session
      description: >
        Clerk session for the single persisted SOY MONEY owner. Human and
        browser operations are owner-only; client-supplied owner identifiers
        are never accepted.
    ServiceHmac:
      type: apiKey
      in: header
      name: X-Service-Signature
      description: >
        HMAC-SHA256 machine authentication. Pair this header with
        X-Service-Id, X-Service-Operation, X-Service-Timestamp, X-Dispatch-Id,
        X-Body-Sha256, and the canonical method/path headers.
...
  schemas:
...
    Opportunity:
      type: object
      required: [id, name, description, sector, problem, targetCustomer, proposedSolution, monetizationMethod, score, estimatedCost, difficulty, risk, timeToRevenue, status, proofStatus, source, sourceUrl, category, titleClaim, evidenceRefs, fingerprint, researchStatus, demandConfidence, detectedAt, validFrom, validUntil, expiresAt, expirationReason, expiredAt, expirationOutcome, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        sector: { type: string }
        problem: { type: string }
        targetCustomer: { type: string }
        proposedSolution: { type: string }
        monetizationMethod: { type: string }
        score: { type: integer }
        estimatedCost: { type: number }
        difficulty: { type: string }
        risk: { type: string }
        timeToRevenue: { type: string }
        status: { type: string }
        proofStatus: { type: string }
        source: { type: string }
        sourceUrl: { type: ["string", "null"], format: uri }
        category: { type: string }
        titleClaim: { type: ["string", "null"] }
        evidenceRefs: { type: array, items: { type: string } }
        fingerprint: { type: ["string", "null"] }
        researchStatus: { type: string }
        demandConfidence: { type: number }
        detectedAt: { type: ["string", "null"], format: date-time }
        validFrom: { type: ["string", "null"], format: date-time }
        validUntil: { type: ["string", "null"], format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        expirationReason: { type: ["string", "null"] }
        expiredAt: { type: ["string", "null"], format: date-time }
        expirationOutcome: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DiscoveryResearchInput:
      type: object
      required: [category]
      properties:
        category:
          type: string
          enum: [BUSINESS, DIGITAL_PRODUCTS, SERVICES, SAAS, AUTOMATION, AFFILIATE, OTHER_LEGAL_OPPORTUNITIES, SPORTS]
        query: { type: string, maxLength: 240 }
        idempotencyKey: { type: string, maxLength: 240 }
    DiscoveryResearchRun:
      type: object
      required: [id, idempotencyKey, category, query, status, sourceCount, independentSourceCount, acceptedCount, scoreBreakdown, startedAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        idempotencyKey: { type: string }
        category: { type: string }
        query: { type: string }
        status: { type: string }
        sourceCount: { type: integer }
        independentSourceCount: { type: integer }
        acceptedCount: { type: integer }
        rejectionReason: { type: ["string", "null"] }
        scoreBreakdown: { type: object, additionalProperties: { type: number } }
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DiscoveryResearchResponse:
      type: object
      required: [run, findings]
      properties:
        run: { $ref: "#/components/schemas/DiscoveryResearchRun" }
        opportunities:
          type: array
          items: { $ref: "#/components/schemas/Opportunity" }
        findings: { type: array, items: { type: object, additionalProperties: true } }
        accepted: { type: boolean }
        error: { type: ["object", "null"], additionalProperties: true }
...