# lib/api-zod/src/generated/types/discoveryResearchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SOY MONEY OS API
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoveryResearchResponseError } from './discoveryResearchResponseError';
import type { DiscoveryResearchResponseFindingsItem } from './discoveryResearchResponseFindingsItem';
import type { DiscoveryResearchRun } from './discoveryResearchRun';
import type { Opportunity } from './opportunity';

export interface DiscoveryResearchResponse {
  run: DiscoveryResearchRun;
  opportunities?: Opportunity[];
  findings: DiscoveryResearchResponseFindingsItem[];
  accepted?: boolean;
  /** @nullable */
  error?: DiscoveryResearchResponseError;
}